ui/requirements.txt
streamlit==1.36.0
requests==2.32.3
python-dotenv==1.0
pandas==2.2
pydantic==2.7
natsort==8.4
fsspec==2024.6
s3fs==2024.6
seaborn==0.13.2